+++ packages/core/src/types/response.ts
@@ -5,6 +5,15 @@ import { InternalRoute } from "./route";
 export type RawParams = { [key: string]: string };
 export type Params = { [key: string]: any };
 
+export interface ModifiableResponseProperties {
+  status?: number;
+  error?: any;
+  body?: any;
+  data?: any;
+  title?: string;
+  redirectTo?: RedirectProps;
+}
+
 export interface GenericResponse {
   location: HickoryLocation;
   key: string;
@@ -25,4 +34,12 @@ export interface Resolved {
   error: any;
   initial: any;
   every: any;
+}
+
+export interface RedirectProps {
+  name: string;
+  params?: Params;
+  hash?: string;
+  query?: any;
+  state?: any;
 }